requirements.txt
PyQt5>=5.15.0
ping3>=4.0.0
psutil>=5.8.0
dnspython>=2.1.0
speedtest-cli>=2.1.0
pysnmp>=4.4.0
python-whois>=0.7.0
